requirements.txt
# Core Dependencies
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
networkx>=3.0

# Data Processing
polars>=0.17.0
dask>=2023.3.0
xarray>=2023.3.0

# Visualization
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0

# API and Web
passlib[bcrypt]
fastapi>=0.95.0
uvicorn>=0.21.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Testing
pytest>=7.3.1
pytest-cov>=4.0.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.17.0
